requirements.txt
aiosmtplib==2.0.2
anyio==3.7.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.4
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-mail==1.3.1
fastapi-utils==0.2.1
greenlet
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.3
passlib==1.7.4
pyasn1==0.5.0
pydantic==1.10.11
PyMySQL==1.1.0
pyodbc
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
pandas
cachetools==5.3.1
openpyxl==3.1.2